refactor: fix some linting issues
