+++ library/src/androidTest/java/ai/bitlabs/sdk/util/ResponseBodyExtensionKtTest.kt
@@ -1,5 +1,6 @@
 package ai.bitlabs.sdk.util
 
+import ai.bitlabs.sdk.util.extensions.body
 import android.util.Log
 import com.google.common.truth.Truth.assertThat
 import okhttp3.MediaType
