--- library/src/core/res/drawable/background_bottom_corners_gradient.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <corners
-        android:bottomLeftRadius="5dp"
-        android:bottomRightRadius="5dp" />
-
-    <gradient
-        android:angle="45"
-        android:endColor="@color/colorPrimaryDark"
-        android:startColor="@color/colorPrimaryDark" />
-</shape>
