+++ library/src/main/java/ai/bitlabs/sdk/data/model/bitlabs/BitLabsResponse.kt
@@ -3,7 +3,7 @@ package ai.bitlabs.sdk.data.model.bitlabs
 import androidx.annotation.Keep
 import com.google.gson.annotations.SerializedName
 
-/** This data class is most likely used in all API repsonses in BitLabs API */
+/** This data class is most likely used in all API responses in BitLabs API */
 @Keep
 internal data class BitLabsResponse<T>(
     val data: T?,
