--- library/src/main/java/ai/bitlabs/sdk/util/BundleExtension.kt
@@ -1,14 +1,0 @@
-package ai.bitlabs.sdk.util
-
-import android.os.Build
-import android.os.Bundle
-import java.io.Serializable
-
-inline fun <reified T : Serializable> Bundle.getSerializableCompat(key: String): T? {
-    return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-        getSerializable(key, T::class.java)
-    } else {
-        @Suppress("DEPRECATION")
-        getSerializable(key) as? T
-    }
-}
