+++ library/src/main/java/ai/bitlabs/sdk/util/extensions/ResponseBodyExtension.kt
@@ -1,6 +1,7 @@
-package ai.bitlabs.sdk.util
+package ai.bitlabs.sdk.util.extensions
 
 import ai.bitlabs.sdk.data.model.bitlabs.BitLabsResponse
+import ai.bitlabs.sdk.util.TAG
 import android.util.Log
 import com.google.gson.GsonBuilder
 import com.google.gson.reflect.TypeToken
