--- library/src/main/res/drawable/ic_sign_out_regular.xml
@@ -1,9 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="512"
-    android:viewportHeight="512">
-  <path
-      android:fillColor="#FFFFFFFF"
-      android:pathData="M96,64h84c6.6,0 12,5.4 12,12v24c0,6.6 -5.4,12 -12,12L96,112c-26.5,0 -48,21.5 -48,48v192c0,26.5 21.5,48 48,48h84c6.6,0 12,5.4 12,12v24c0,6.6 -5.4,12 -12,12L96,448c-53,0 -96,-43 -96,-96L0,160c0,-53 43,-96 96,-96zM327.1,83.5l-19.6,19.6c-4.8,4.8 -4.7,12.5 0.2,17.1L420.8,230L172,230c-6.6,0 -12,5.4 -12,12v28c0,6.6 5.4,12 12,12h248.8L307.7,391.7c-4.8,4.7 -4.9,12.4 -0.2,17.1l19.6,19.6c4.7,4.7 12.3,4.7 17,0l164.4,-164c4.7,-4.7 4.7,-12.3 0,-17l-164.4,-164c-4.7,-4.6 -12.3,-4.6 -17,0.1z"/>
-</vector>
